requirements.txt
blinker==1.6.2
certifi==2023.7.22
chardet==3.0.4
click==8.1.6
Flask==2.3.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
pdfminer==20191125
pycryptodome==3.18.0
python-docx==0.8.11
rfc3986==1.5.0
sniffio==1.3.0
Werkzeug==2.3.6
zipp==3.16.2
